refactor(RuntimeUpdates): decouple script runtime from a generic runtime
This will allow for programmatic execution of arbitrary functions without having to rely on scripts.
