+++ src/uipath/_cli/cli_eval.py
@@ -13,7 +13,7 @@ from uipath._cli._runtime._contracts import (
     UiPathRuntimeContextBuilder,
     UiPathRuntimeFactory,
 )
-from uipath._cli._runtime._runtime import UiPathRuntime
+from uipath._cli._runtime._runtime import UiPathScriptRuntime
 from uipath._cli.middlewares import MiddlewareResult, Middlewares
 
 from .._utils.constants import ENV_JOB_ID
@@ -62,11 +62,12 @@ def eval_agent_middleware(
         )
 
     try:
-        runtime_factory = UiPathRuntimeFactory(UiPathRuntime, UiPathRuntimeContext)
+        runtime_factory = UiPathRuntimeFactory(
+            UiPathScriptRuntime, UiPathRuntimeContext
+        )
         context = (
             UiPathRuntimeContextBuilder()
             .with_defaults(**kwargs)
-            .with_entrypoint(entrypoint)
             .with_entrypoint(entrypoint)
             .mark_eval_run()
             .build()
